requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
attrs==26.1.0
bcrypt==5.0.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cohere==5.21.1
colorama==0.4.6
cryptography==47.0.0
dataclasses-json==0.6.7
datasets==4.8.5
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
fastavro==1.12.2
filelock==3.29.0
filetype==1.2.0
FlashRank==0.2.10
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.50.0
google-genai==1.74.0
greenlet==3.5.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.13.0
idna==3.13
instructor==1.15.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.17
langchain-classic==1.0.4
langchain-cohere==0.5.1
langchain-community==0.4.1
langchain-core==1.3.2
langchain-google-genai==4.2.2
langchain-huggingface==1.2.2
langchain-openai==1.2.1
langchain-pinecone==0.2.13
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.13
langsmith==0.8.0
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
-e git+https://github.com/yashveersaini/RAG-Based-Medical-AI-assistant.git@cae3ebe3626d0e0a1dffefb60e1a70b401d25c34#egg=medical_assistant
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
onnxruntime==1.25.1
openai==2.33.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
pandas==3.0.2
pillow==12.2.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
propcache==0.4.1
protobuf==7.34.1
psycopg2-binary==2.9.12
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
PyMuPDF==1.27.2.3
pypdf==6.10.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
ragas==0.4.3
rank-bm25==0.2.2
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.4
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.7.0
typer==0.25.1
types-PyYAML==6.0.12.20260408
types-requests==2.33.0.20260408
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
